- fixing the price to be double

+++ e-store/e-store/ScenesScreen/CategoryItems/Model/CategoryItem.swift
@@ -12,11 +12,11 @@ import Kingfisher
 class CategoryItem {
     var imgURL: String
     var name: String
-    var price: Int
+    var price: Double
     var rating: Float
     var reviews: Int
     
-    init(img: String, name: String, price: Int, rating: Float, reviews: Int) {
+    init(img: String, name: String, price: Double, rating: Float, reviews: Int) {
         
         self.imgURL = img
         self.name = name
